exp-server/requirements.txt
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
python-jose==3.4.0
gunicorn==21.2.0

celery==5.5.3
redis==4.5.5

SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
pg8000==1.30.5

PyJWT==2.10.1
bcrypt==4.1.2
cryptography==42.0.5
python-dotenv==1.0.0